n[nting

+++ Yakuza_Syndicate/Yakuza_Syndicate/GangMembers.h
@@ -7,7 +7,7 @@ private:
 	sf::Font font;
 
 	int amount;
-	float upkeep;
+	int upkeep;
 	bool action;
 	bool inFriendlyTerr;
